fix: improved tasks and modified .devcontainer
