+++ .devcontainer/Dockerfile
@@ -14,9 +14,9 @@ ENV HOME="/root"
 # --------------------------------------
 # Git
 # --------------------------------------
-# Need to add the devcontainer workspace folder as a safe directory to enable git 
+# Need to add the devcontainer workspace folder as a safe directory to enable git
 # version control system to be enabled in the containers file system.
-RUN git config --global --add safe.directory "/workspaces/plugin-template"
+RUN git config --global --add safe.directory "/workspaces/plugin-microsoft365"
 # --------------------------------------
 
 # --------------------------------------
@@ -53,11 +53,11 @@ ENV PATH="$PATH:$JAVA_HOME/bin"
 # Will load a custom configuration file for Micronaut
 ENV MICRONAUT_ENVIRONMENTS=local,override
 # Sets the path where you save plugins as Jar and is loaded during the startup process
-ENV KESTRA_PLUGINS_PATH="/workspaces/plugin-template/local/plugins"
+ENV KESTRA_PLUGINS_PATH="/workspaces/plugin-microsoft365/local/plugins"
 # --------------------------------------
 
 # --------------------------------------
-# SSH 
+# SSH
 # --------------------------------------
 RUN mkdir -p ~/.ssh
 RUN touch ~/.ssh/config
