+++ src/main/java/io/kestra/plugin/microsoft365/outlook/AbstractMicrosoftGraphIdentityConnection.java
@@ -1,11 +1,10 @@
-package io.kestra.plugin.microsoft365;
+package io.kestra.plugin.microsoft365.outlook;
 
 import com.azure.identity.ClientSecretCredential;
 import com.azure.identity.ClientSecretCredentialBuilder;
 import com.microsoft.graph.serviceclient.GraphServiceClient;
 import io.kestra.core.models.property.Property;
-import io.kestra.core.models.triggers.AbstractTrigger;
-import io.kestra.core.models.triggers.PollingTriggerInterface;
+import io.kestra.core.models.tasks.Task;
 import io.kestra.core.runners.RunContext;
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -13,7 +12,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.time.Duration;
 import java.util.Optional;
 
 @SuperBuilder
@@ -21,7 +19,7 @@ import java.util.Optional;
 @EqualsAndHashCode
 @Getter
 @NoArgsConstructor
-public abstract class AbstractMicrosoftGraphIdentityPollingTrigger extends AbstractTrigger implements PollingTriggerInterface {
+public abstract class AbstractMicrosoftGraphIdentityConnection extends Task {
     @Schema(title = "Tenant ID", description = "Azure tenant ID (directory ID)")
     @NotNull
     protected Property<String> tenantId;
@@ -41,10 +39,6 @@ public abstract class AbstractMicrosoftGraphIdentityPollingTrigger extends Abstr
     @Builder.Default
     protected Property<String> scopes = Property.ofValue("https://graph.microsoft.com/.default");
 
-    @Schema(title = "Polling interval", description = "Interval between polls")
-    @Builder.Default
-    protected Duration interval = Duration.ofMinutes(5);
-
     protected GraphServiceClient createGraphClient(RunContext runContext) throws Exception {
         String rTenantId = runContext.render(tenantId).as(String.class).orElseThrow();
         String rClientId = runContext.render(clientId).as(String.class).orElseThrow();
@@ -62,10 +56,5 @@ public abstract class AbstractMicrosoftGraphIdentityPollingTrigger extends Abstr
     protected Optional<String> getUserPrincipalName(RunContext runContext) throws Exception {
         if (userPrincipalName == null) return Optional.empty();
         return runContext.render(userPrincipalName).as(String.class);
-    }
-
-    @Override
-    public Duration getInterval() {
-        return this.interval;
     }
 }
